requirements.txt
openai
SpeechRecognition
google-cloud-texttospeech
playsound
python-dotenv
dlib
keras-facenet
scipy
annoy
google-cloud-speech
psycopg2-binary
tensorflow
numpy
pyaudio